Fix book deletion and add modal styling
Book delete button was using the index for book creation, not the index
of the target book itself. Therefore, the event listener was moved down
and merged with read status toggle's event listener so that both
functionalities can share the same target book index variable.

Styled the add new book modal to make it consistent with the rest of
the project.